Feat: MyPage 관련 뷰 폴더링, 모델, API 분리
- MyPage 관련 뷰 폴더링
- 모델, API 분리
