+++ Ting/Models/UserInfo.swift
@@ -19,26 +19,3 @@ struct UserInfo: Identifiable, Codable {
     let location: String
     let interest: String
 }
-
-// Create
-class UserInfoService {
-    static let shared = UserInfoService()
-    private let db = Firestore.firestore()
-    private init() {}
-    
-    func createUserInfo(info: UserInfo, completion: @escaping (Result<Void, Error>) -> Void) {
-        do {
-            let data = try Firestore.Encoder().encode(info) // Firestore에 저장할 수 있는 형태로 인코딩
-            db.collection("infos").addDocument(data: data) { error in
-                if let error = error {
-                    completion(.failure(error))
-                } else {
-                    completion(.success(()))
-                }
-            }
-        } catch {
-            completion(.failure(error))
-        }
-    }
-    
-}
