+++ Ting/MyPage/DeleteInfo/DeleteInfoVC.swift
